addressed review changes (except nav bar color + logo)
