+++ public/locales/en/menu.json
@@ -1,5 +1,5 @@
 {
-    "home": "Home",
+    "map": "Map",
     "about": "About",
     "healthInfo": "Health Information",
     "admin": "Admin",
