+++ public/locales/es/home.json
@@ -1,5 +1,4 @@
 {
-    "heading": "Tablero de la calidad del aire en South Gate",
     "noSensorGauge": "¡Haz clic en un sensor en el mapa para ver el Índice de la calidad de aire (Air Quality Index, AQI) aquí!",
     "dataVizBox": "Cuadro de visualización de datos",
     "expandMap": "Expandir para ver el mapa de sensores",
