+++ public/locales/es/menu.json
@@ -1,5 +1,5 @@
 {
-    "home": "Inicio",
+    "map": "Mapa",
     "about": "Sobre Nosotros",
     "healthInfo": "Información de Salud",
     "admin": "Administración",
